added some styling (incomplete)
