+++ scranShipper/src/RestaurantMenu/style.css
@@ -14,8 +14,10 @@
     align-items: flex-start;
     justify-content: flex-start;
     margin: 0.5rem;
-    padding: 0.5rem;
-    width: 15%;
+    padding: 1rem;
+    width: 18%;
+    height: 15rem;
+    border: 1px solid black;
 }
 
 .name {
@@ -27,6 +29,7 @@
     flex-direction: row;
     flex-wrap: wrap;
     font-size: 1rem;
+    color: white;
 }
 
 .calories {
@@ -37,14 +40,39 @@
     background-color: #1e90ff;
 }
 
-.side {
+.isSide {
     background-color: #ffcc00;
 }
 
-.price {
-
+.isBreakfast {
+    background-color: #8b0000;
 }
 
-button {
+.price {
+    font-size: 1rem;
+}
 
+.bottom {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.button {
+    background-color: #1e90ff;
+    width: 40px;
+    height: 40px;
+    color: white;
+    margin: 0 1rem;
+}
+
+.buttonsAndQuantity {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+    align-items: flex-start;
+    width: 100%;
 }
